fix: address three correctness issues from Opus review
- parseInboundBuffer: cast bounds sum to uint64_t to prevent uint32
  overflow with adversarial dataOfs+dataLen values
- rndisCommand: clear responseAvailable and arm interrupt pipe BEFORE
  SEND_ENCAPSULATED_COMMAND to close the race where a fast device
  delivers RESPONSE_AVAILABLE before the clear, causing a 1 s timeout
- openInterruptPipe: release intPipe/intBuf on each partial failure
  path so a non-fatal failure doesn't leave half-initialized state
  that rndisCommand would then block on waiting for a notification
- tests: add oversizedDataOfsRejected and nearOverflowSumRejected cases
  to cover the two previously untested bounds-check branches (15/15)

+++ TetherKitDriver/RNDISProtocolCore.cpp
@@ -28,7 +28,7 @@ ParseResult parseInboundBuffer(const uint8_t * buf,
         }
         if (dataOfs > msgLen ||
             dataLen > msgLen ||
-            (8u + dataOfs + dataLen) > msgLen) {
+            ((uint64_t)8u + dataOfs + dataLen) > (uint64_t)msgLen) {
             return ParseResult::INVALID_DATA_BOUNDS;
         }
         if (dataLen > 0 && cb) {
